Moves Blog to Index
