+++ projects/frontend/src/data/about.md
@@ -1,0 +1,5 @@
+I’m a full-stack web developer work at [determined.ai](https://determined.ai)
+in San Francisco, CA. Our platform enables users to build and collaborate on
+A.I. based applications faster. I've been tinkering with the web since 2013 and
+have worked for startups and enterprises alike since then. Reach out and
+definitely check out my blog!
